Leitura de midias enchaminhada para o sucesso

+++ Trabalho/codes/Musica.cpp
@@ -3,12 +3,12 @@
 Musica::Musica(){
     cout << "Musica Publicada"<<endl;
 }
-Musica::Musica(string nome, Genero gen, int duracao, int anoLancamento): Midia(nome, duracao, genero){
+Musica::Musica(string nome, Genero gen, float duracao, int anoLancamento): Midia(nome, duracao, genero){
     this->nome = nome;
     this->duracao = duracao;
     this->anoLancamento = anoLancamento;
     this->genero = gen;
 }
-// void Musica::imprimeinfoProduto();
+// void Musica::imprimeinfoMidia();
 // void Musica::imprimeNoArquivo(ofstream &outfile);
 // void Musica::formataDuracao();
